Move the Contact* files inside Contacts component
