+++ src/Contacts/CMakeLists.txt
@@ -5,14 +5,13 @@
 
 if (FRAMEWORK_COMPILE_Contact)
   set(H_PREFIX include/BipedalLocomotion/Contacts)
-
   add_bipedal_locomotion_library(
     NAME                  Contacts
-    IS_INTERFACE
-    PUBLIC_HEADERS        ${H_PREFIX}/Contact.h 
+    PUBLIC_HEADERS        ${H_PREFIX}/Contact.h ${H_PREFIX}/ContactList.h ${H_PREFIX}/ContactPhase.h ${H_PREFIX}/ContactPhaseList.h
+    SOURCES               src/Contact.cpp src/ContactList.cpp src/ContactPhase.cpp src/ContactPhaseList.cpp
     PUBLIC_LINK_LIBRARIES MANIF::manif
     INSTALLATION_FOLDER   Contacts)
 
+    add_subdirectory(tests)
+
 endif()
-
-
