+++ src/Contacts/src/ContactPhase.cpp
@@ -5,13 +5,12 @@
  * distributed under the terms of the GNU Lesser General Public License v2.1 or any later version.
  */
 
-#include <BipedalLocomotion/Planners/ContactPhase.h>
+#include <BipedalLocomotion/Contacts/ContactPhase.h>
 #include <algorithm>
 
-using namespace BipedalLocomotion::Planners;
+using namespace BipedalLocomotion::Contacts;
 
 bool ContactPhase::isListIncluded(const std::string &key) const
 {
     return  activeContacts.find(key) != activeContacts.end();
 }
-
