+++ src/Contacts/tests/CMakeLists.txt
@@ -1,0 +1,15 @@
+# Copyright (C) 2020 Istituto Italiano di Tecnologia (IIT). All rights reserved.
+# This software may be modified and distributed under the terms of the
+# GNU Lesser General Public License v2.1 or any later version.
+
+add_bipedal_test(
+    NAME ContactList
+    SOURCES ContactListTest.cpp
+    LINKS BipedalLocomotion::Contacts Eigen3::Eigen)
+
+add_bipedal_test(
+  NAME ContactPhaseList
+  SOURCES ContactPhaseListTest.cpp
+  LINKS BipedalLocomotion::Contacts)
+
+
